requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
psycopg2-binary==2.9.6
sqlalchemy==2.0.15
alembic==1.11.2
pydantic[email]==1.10.7
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
pytest==7.3.1
pytest-asyncio==0.21.0
httpx==0.24.1
python-multipart==0.0.6
redis==4.6.0

# Problem 3 async / benchmarking deps
asyncpg==0.28.0
pytest-benchmark==4.0.0

# Monitoring / Metrics
prometheus-fastapi-instrumentator==6.1.0
prometheus-client==0.17.1

# Coverage
pytest-cov==4.1.0
